requirements.txt
fastapi[standard]
pymongo[srv]==4.3.3
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
email-validator==2.0.0.post2